requirements.txt
django~=3.2.5
requests~=2.26.0
djangorestframework~=3.13.1